nimbase: add VxWorks platform specific includes
Some VxWorks system includes are not self contained and need to be
included in a particular order because they depend on some other
include files (otherwise the compilation fails). We add those include
files that must be included before the rest to nimbase.h to ensure that
they are included first.