Removing -fasmblocks default for OS X
Considering that alternatives to the Apple compiler versions of xcode like GCC 4.9 (as installed by Homebrew) error out because it is an unknown parameter. Modern Xcode's clang + gcc do not seem to need this anymore. 

I think it is ok to remove because developer with old xcode are rare and can easily add this to their systems if needed.