Improve website background quality
The old image was compressed with JPG and was improperly rendered anyway. The
background has been modified so that each dot lines up perfectly on pixel
boundaries. It's compressed with lossy PNG, which has the advantage of
preserving the quality of the dots.

There will be another PR to fix the background URL for the forum.