Add additional information to mapType internalError.
I wrote this bad program:

 $ cat nil_crash_compiler.nim
 # Crashes compiler
 type
   O = object
     i: int

 var
   o: O

 o = cast[O](nil)

And this causes the compiler to crash with very little information:

  Error: internal error: mapType

This adds additional information to the internal error and now it outputs:

  Error: internal error: mapType: could not map 'typ:tyNil' to a TCTypeKind

I suggest that mapType should probably return a result for tyNil and then
the compiler should probably produce a better error message.

But in the meantime adding more information can be helpful.