Add access to thread handle under Windows
Under Windows, createThread returns the thread handle, which is placed in GcThread.sys - this is of type SysHandle (itself an alias of Handle under Windows).
This PR simply exposes a function "threadHandle" to return the handle so that it may be used in Windows.

Because this function is only exposed under Windows, I chose to set the return type as Handle, rather than SysThread (which is not an exposed type).
If anyone feels the SysThread should be exposed and the function should return this type instead, just comment and I'll change it. I haven't done any developing under Linux, but a quick google implies that GcThread.sys is also set to be a thread handle in this OS. If SysThread is exposed then the "threadHandle" function might as well exist under other operating systems too.