Removed libiconv prefix (on OS X) to make encodings compile on OS X.
I am not sure since which version OS X does include the libiconv library
without the `lib` prefix. But it seems this is the case for some years
now. If there are ways to check for the OS X version at compile time
this may be needed to support older OS X versions. But I guess thats not
needed for most users working with Nim anyway.