Add 'auto-decoration' to nimGetProcAddr
Maintainers of win32 DLLs sometimes opt to put "decorated" function
names into the DLL's symbol table.  (Google "stdcall name
decoration").  To pull a function pointer out of a DLL like this, you
have to provide the decorated version of the function's name.  This is
painful for the authors of NIM wrappers: they have to manually add
weird punctuation to their 'importc' strings, and to make matters
worse, they have to do this conditionally only on win32.

This commit adds 'auto-decoration' to nimGetProcAddr.  This function
probes the DLL for the undecorated name, and if it doesn't find it, it
adds decoration and tries again.  This makes it so that the author of
the wrapper doesn't have to worry about it.