Fix the field values to be null when the release schema is compiled.
Fix the field values to be null when the release schema is compiled.
tempcCnt,A field to store the number of temporary variables, for unknown reasons, 
after performing a sqlgetdata function and circulating variables cCnt value will be changed to 0,
so the values of the temporary variable to store the cCnt. 
After every cycle and specified to cCnt. To ensure the traversal of all fields.