openssl: fix compatibility with OpenSSL 1.0.2
OpenSSL has removed SSLv3 support in 1.0.2 and thus `SSLv3_method()` is
no longer available. If an SSL-enabled Nim program runs on a system with
OpenSSL 1.0.2, an error `could not import SSLv3_method` will occur.

This commit adds a flag `nosslv3`. When enabled, the `SSLv3_method()`
will be replaced with a stub function, and thus it will run properly
with OpenSSL 1.0.2 while not breaking any other module. Of course, the
modules will not be able to connect to SSLv3 servers with this enabled.