Added `.*/`, `~*` and `*.log` to GIT ignore list.
Using differend IDEs for the Nim compiler I am constantly running
into the problem that my global .gitignore is not working for the Nim
directory. This is caused by the `*` / `!*.*` like rules, which
reset all other rules. So I thought a bit about a sensible filtering
of common IDE special files and added that to the .gitignore.