Fix #4475
Existing implementation would append the default value for a varargs parameter (the empty array) to the end of the sons of the
nnkCall node, rather than storing it into the correct index. This left the location where it should have been stored set to
its default value of nil, causing later code that uses this node to segfault.