Nicer English
Line 1264: A sequence may be passed to an openarray parameter.
This is not explained until a bit further on so should be deleted or should provide a cross-ref for openarray.

Line 1389: b[11..^2] = "useful"
This deserves some explanation. I'm guessing "Replace in be from position 11 to length-2"?

I see you are using British (European) English, e.g., "behaviour"; that's nice.

Line 1576: proc `*` *(a, b: seq[int]): seq[int] =
My guess is that the `*` is the function name and the second * is marking it for export. Seems strange that the export * isn't after the return type (e.g., to be like var x*: int). But anyway, given the use of two *s I think you should at least say explictly which is which in the text that follows it.

Line 1580: for i in 0..len(a)-1: result[i] = a[i] * b[i]
Why isn't this: for i in 0..<len(a): ... 

I'd have put the excluding symbols section at the end of the imports section since it is unusual and probably rare.