Have the linker search `/usr/local/lib` on Linux
This was causing some issues when I was trying to link a `.so` of `raylib` and `glfw`.  `/usr/local/lib` wasn't in the `-rpath` option at link time on Linux.  It is a common location for users to install libraries on Linux and should be part of the link path.