parsecsv.open on empty files
parsecsv.error was using the CsvParser.filename for formatting the error
message - this is problematic, as if the file can't be opened, this will
be nil and will cause a SIGSEGV when handling the original exception
(can't open the file).

This fix uses a default empty string if the CsvParser.filename is nil.