Add ByteStream and generic `read` proc to streams
This adds a ByteStream that reads and writes `seq[byte]` to and from a
stream. Might be useful for cryptography, or serialization.

Along with that it also adds a generic `read` proc that takes the type
you want to read instead of the old pattern of `read<typeName>`. This
means that things like macros can easier create read statements, and it
means that type aliases can be used as well like this `type MyInt = int16;
stream.read(MyInt)`. The old read procs have been marked as deprecated,
however since these are used in so many different libraries around it
would be helpful to after a while put them behind a compile-time switch
instead of removing them outright.