small changes to editDistance
the first change prevents checking the null byte (which is not secure when it occurs inside the string and should be changed anyways).

the second change removes a tautologic check:
at line 1905 len1 is still always <= len2 because the same operations where done to both and at the begin len1 is also definetly <= len2.
Especially this means that if `len1 > 0` also `len2 > 0` is always true.