Fix isUpperAscii, isLowerAscii for strings with non-alpha chars
This commit changes the behavior of isUpperAscii and isLowerAscii in
the strutils module -- only for the case where the input is a *string*
with a mix of alphabetic and non-alphabetic characters.

The new behavior mimics the Python isupper and islower behavior
i.e. non-alphabetic chars are ignored when checking if the whole
string is upper-case or lower-case.

    Before: doAssert(not "A B".isUpperAscii) passed
    Now:    doAssert "A B".isUpperAscii passes

.. and the similar for isLowerAscii.

To get the old behavior (the old behavior checked if the whole string
contained only alphabetic characters AND if all characters were
upper/lower case), simply do:

    str.isAlphaAscii and str.isUpperAscii
    str.isAlphaAscii and str.isLowerAscii

(where str is a variable of type string)

See https://github.com/nim-lang/Nim/issues/7963 for more info.