Fix isUpper, isLower for strings with non-alpha unicode chars
Fixes https://github.com/nim-lang/Nim/issues/7963.

This commit changes the behavior of isUpper and isLower in the
unicode module -- only for the case where the input is a *string*
with a mix of alphabetic and non-alphabetic unicode characters.

The new behavior mimics the Python isupper and islower behavior
i.e. non-alphabetic chars are ignored when checking if the whole
string is upper-case or lower-case.

    Before: doAssert(not "A B".isUpper) passed
    Now:    doAssert "A B".isUpper passes

.. and the similar for isLower.

To get the old behavior (the old behavior checked if the whole string
contained only alphabetic unicode characters AND if all characters
were upper/lower case), simply do:

    str.isAlpha and str.isUpper
    str.isAlpha and str.isLower

(where str is a variable of type string)