unittest: opt-in test parallelisation
Run tests in parallel, using threadpool's "spawn", when
"--threads:on -d:nimtestParallel" is passed to the compiler.

The tests need to be GC-safe, for this to work, and any enabled output
formatters need to be able to be initialised multiple times in multiple
threads - which excludes JUnitOutputFormatter as it is right now because
it writes to the stream each time it's instantiated.