parseopt: better documentation
Greatly expand on the top-level documentation and elaborate on
details that may have been ambiguous or unclear, especially
regarding the shortNoVal and longNoVal parameters. Modify existing
proc descriptions and add "See also" links as well.

Also, add examples to procs that do not have them and tweak the
existing examples for the getopt iterators so that they compile
without any changes.