fix `customPragmaNode` call for certain objects
I'm a little puzzled about this. With the change introduced in the
previous commit to add the `noserialize` option for JSON, the first
example in the `tjsonmacro.nim` does not work anymore. The generic
`Point[float]` cannot be serialized, because of a `node is not a
symbol` error. In order to get the type def implementation under the
given case, we have to also get the implementation of the underlying
type.

Maybe this should be fixed in a different way?