Add skipDirs set to os.walkDirRec iterator
Sometimes when recursively traversing a directory tree it is necessary
to skip some of the sub directories. If the filtering is applied outside
of the iterator, then it will be inefficient because the iterator will
still recurse in the unwanted directory sub-tree. To avoid this
additional HashSet[string] parameter is added to the os.walkDirRec
iterator which contains a list of relative directory paths for the
directories which must be skipped.

Example:
for file in walkDirRec(".", skipDirs = toHashSet([".git", ".hg"])):
  echo file