Add channels example
This PR revives #8611 since the channels module still does not have an
example of usage. I attempted to fix the outstanding issue that caused
the previous PR to be closed, namely implementing the example on top of
threads instead of threadpool. I also added an additional example of
using sharedAlloc0 to create channel pointers based on the discussion in
 #4748. Please let me know if there is anything that needs to be
changed!