New implementation that use a interator instance of future streams
The previous suggestion don't work with future streams because all data is write to the the future stream and used later, but only when the queue is full and the future is complete as Dominik Picheta mentioned in the comment. This new version uses an iterator to read the chunks of data.