net: the complete SIGPIPE avoidance package for SSL_shutdown()
We are obligated to reply to a close_notify notification, but then the
peer is also allowed to just disconnect from us once they sent that
message. This can cause SIGPIPE to be raised on socket closing as we
attempts to do a graceful shutdown via SSL_shutdown().

This commit does the following:
- Ignore SIGPIPE while we are doing a shutdown.
- If we received a shutdown notification from the peer, ignore
  disconnection errors that might be raised as the peer might have
  already closed their side of the connection.