typetraits: add rangeof(T), a shortcut for low(T)..high(T)
This func returns a slice with the full range of type `T`, basically a
shortcut for `low(T)..high(T)`. This is useful for verifying whether a
value of one type can be converted to an another.

The main use case for this would be as a sugar for the initial
conversion between an ordinal and a `range` type, though it's flexible
enough that it can be used for many other situations, like conversions
between int/uint of different sizes.