Update documentation.
I have included a little note which would help clarify the requirements for the use of the db_mysql and mysql modules. Namely, the requirement for libmysql, and the issues which may arise on systems which do not include a mysql shell installation (note: these experiences were made on Windows 10).